Show size of deletions and large insertions
Add a label to deletions and large insertions to show the size of the variant.
For deletions, show the size centered over the "gap line" as long as the size
label fits in the available gap.  For large insertions, show a label with the
size in place of the "large insertion" red crosshairs.